+++ packages/bsky/src/lexicon/types/com/atproto/server/getSession.ts
@@ -16,6 +16,7 @@ export interface OutputSchema {
   handle: string
   did: string
   email?: string
+  emailConfirmed?: boolean
   [k: string]: unknown
 }