+++ packages/dev-env/src/util.ts
@@ -44,6 +44,16 @@ export const mockResolvers = (idResolver: IdResolver, pds: TestPds) => {
   }
 }
 
+export const mockMailer = (pds: TestPds) => {
+  const mailer = pds.ctx.mailer
+  const _origSendMail = mailer.transporter.sendMail
+  mailer.transporter.sendMail = async (opts) => {
+    const result = await _origSendMail.call(mailer.transporter, opts)
+    console.log(`✉️ Email: ${JSON.stringify(result, null, 2)}`)
+    return result
+  }
+}
+
 const usedLockIds = new Set()
 export const uniqueLockId = () => {
   let lockId: number