+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atproto/dev-env
 
+## 0.2.11
+
+### Patch Changes
+
+- [#1568](https://github.com/bluesky-social/atproto/pull/1568) [`41ee177f`](https://github.com/bluesky-social/atproto/commit/41ee177f5a440490280d17acd8a89bcddaffb23b) Thanks [@dholms](https://github.com/dholms)! - Added email verification and update flows
+
+- Updated dependencies [[`41ee177f`](https://github.com/bluesky-social/atproto/commit/41ee177f5a440490280d17acd8a89bcddaffb23b), [`41ee177f`](https://github.com/bluesky-social/atproto/commit/41ee177f5a440490280d17acd8a89bcddaffb23b)]:
+  - @atproto/api@0.6.20
+  - @atproto/pds@0.1.20
+  - @atproto/common-web@0.2.1
+  - @atproto/bsky@0.0.11
+  - @atproto/identity@0.2.1
+  - @atproto/syntax@0.1.2
+  - @atproto/xrpc-server@0.3.2
+
 ## 0.2.10
 
 ### Patch Changes